Make project into a cement app.
This will help to add support for command line arguements. Also, the relevant directories for building the kernel such as the toolchain directory and export directory can be handled by the cement framework. TO complete this move, the main modules of the project have been moved from kernel_builder to kbuilder/core.