kbuilder/core: remove info() from messages.py
info prints out messages in blue that inform the user on is
happening. However, such messages can and should be printed in white
since they are not as important as other types of information such as if
compilation is successful or not. Removing this also helps to clean up
the code.